[TEC-3473] Rename Flow::Order to FlowcommerceSpree::OrderSync service
Fix flow's Spree::Order decorator:

-`flow_order` method to not fail on nil flow_data

- 'flow_total` method to correctly call format_default_price from FlowcommerceSpree::Api
